feat: implement AuthorisedRESTRequestHandler for enhanced authorization support
